+++ examples/contracts/ERC20/ERC20.rs
@@ -3,9 +3,9 @@
 
 #[brush::contract]
 pub mod erc_20 {
-	use ink_storage::Mapping;
-	use ink::prelude::string::String;
 	use brush::traits::AccountId;
+	use ink::prelude::string::String;
+	use ink_storage::Mapping;
 
 	#[ink(storage)]
 	#[derive(Default, SpreadAllocate)]
